Ajustes globales: Cambios DAO, SQL, Probado base de datos y sincronización entre Cat y Prod. Añadido Mem y File repositorys movil funciona

+++ composeApp/src/commonMain/kotlin/ies/sequeros/com/dam/pmdm/administrador/aplicacion/productos/crear/CrearProductoUseCase.kt
@@ -21,7 +21,7 @@ class CrearProductoUseCase(private val repositorio: IProductoRepositorio, privat
             pendienteEntrega = true,
             idCategoria = crearProductoCommand.idCategoria,
             descripcion = crearProductoCommand.descripcion,
-            precio = crearProductoCommand.precio,
+            precio = crearProductoCommand.precio.toFloat(),
             activar = crearProductoCommand.activar
         )
         val element = repositorio.findByName(item.nombre)
